Module creation
I separated the code for the client and server creation from the
simulation. The goal is to be able to run several simulations and not
only one.
